🚚 [RUM] rename file viewTracker to viewCollection
The view logic will be changed to emit lifeCycle events. Like other
component emiting lifeCycle events, name the file as "xxCollection"
